requirements.txt
openai
requests
beautifulsoup4
azure-identity
azure-core
python-dotenv
prompt_toolkit
rich